Add docs and some minor refactorings
