+++ FaunaDB.Client/FaunaDB.Client.csproj
@@ -87,7 +87,7 @@
     <Compile Include="Query\PageHelper.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Types\BytesV.cs" />
-    <Compile Include="Encoding\Encoder.cs" />
-    <Compile Include="Attributes\FaunaDBAttributes.cs" />
+    <Compile Include="Types\Encoder.cs" />
+    <Compile Include="Types\FaunaDBAttributes.cs" />
   </ItemGroup>
 </Project>
